+++ prepare_release.sh
@@ -1,0 +1,3 @@
+poetry version $1
+dephell deps convert
+poetry run black setup.py