+++ pyquil/__init__.py
@@ -1,3 +1,5 @@
-from pyquil.version import __version__
+from pyquil._version import pyquil_version
 from pyquil.quil import Program
 from pyquil.api import list_quantum_computers, get_qc
+
+__version__ = pyquil_version