+++ pyquil/experiment/tests/test_setting.py
@@ -27,9 +27,7 @@ def _generate_random_states(n_qubits, n_terms):
     all_s_inds = np.random.randint(len(oneq_states), size=(n_terms, n_qubits))
     states = []
     for s_inds in all_s_inds:
-        state = functools.reduce(
-            mul, (oneq_states[pi](i) for i, pi in enumerate(s_inds)), TensorProductState([])
-        )
+        state = functools.reduce(mul, (oneq_states[pi](i) for i, pi in enumerate(s_inds)), TensorProductState([]))
         states += [state]
     return states