+++ pyquil/simulation/tests/test_tools.py
@@ -400,9 +400,7 @@ def test_lifted_state_operator():
     proj_one = one @ one.conj().T
     assert proj_one.shape == (2, 2)
 
-    np.testing.assert_allclose(
-        np.kron(proj_one, proj_plus), lifted_state_operator(xz_state, qubits=[5, 6])
-    )
+    np.testing.assert_allclose(np.kron(proj_one, proj_plus), lifted_state_operator(xz_state, qubits=[5, 6]))
 
 
 def test_lifted_state_operator_backwards_qubits():
@@ -417,9 +415,7 @@ def test_lifted_state_operator_backwards_qubits():
     proj_one = one @ one.conj().T
     assert proj_one.shape == (2, 2)
 
-    np.testing.assert_allclose(
-        np.kron(proj_plus, proj_one), lifted_state_operator(xz_state, qubits=[6, 5])
-    )
+    np.testing.assert_allclose(np.kron(proj_plus, proj_one), lifted_state_operator(xz_state, qubits=[6, 5]))
 
 
 def test_lifted_gate_with_nonconstant_params():